Added the beginnings of virtual interval timer support for MX.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@325 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_interval_timer.h
@@ -10,7 +10,7 @@
  *
  *----------------------------------------------------------------------
  *
- * Copyright 2004 Illinois Institute of Technology
+ * Copyright 2004, 2006 Illinois Institute of Technology
  *
  * See the file "LICENSE" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
@@ -28,7 +28,7 @@
 struct mx_interval_timer_struct {
 	int timer_type;
 	double timer_period;
-	int num_overruns;
+	unsigned long num_overruns;
 	void (*callback_function)( struct mx_interval_timer_struct *itimer,
 				void *callback_args );
 	void *callback_args;
